pyproject.toml
[project]
name = "telegramshop"
version = "0.1.0"
description = ""
authors = [
    {name = "Satocu",email = "hojicimo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.15.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
telegramshop = "telegramshop.main:run"

[tool.poetry]
packages = [{include = "telegramshop", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
